[refactor] 리포지토리 코드 리팩토링
