+++ src/main/java/com/example/aniwhere/repository/RefreshTokenRepository.java
@@ -1,13 +1,10 @@
-package com.example.aniwhere.infrastructure.persistence;
+package com.example.aniwhere.repository;
 
 import com.example.aniwhere.domain.token.RefreshToken;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
-@Repository
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
 	Optional<RefreshToken> findByUserId(Long userId);
-	Optional<RefreshToken> findByRefreshToken(String refreshToken);
 }
