+++ src/main/java/com/example/aniwhere/repository/UserRepository.java
@@ -1,12 +1,10 @@
-package com.example.aniwhere.infrastructure.persistence;
+package com.example.aniwhere.repository;
 
 import com.example.aniwhere.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
-@Repository
 public interface UserRepository extends JpaRepository<User, Long> {
 	Optional<User> findByEmail(String email);
 }
